worker/requirements.txt
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
beautifulsoup4==4.12.2
lxml==4.9.4
ics==0.7.2
feedparser==6.0.10
python-dateutil==2.8.2
requests==2.31.0
